feat : 엔티티 수정

+++ chat-da/src/main/java/site/chatda/domain/school/entity/ClassesId.java
@@ -18,10 +18,10 @@ public class ClassesId implements Serializable {
     private Integer schoolId;
 
     @NotNull
-    @Column(columnDefinition = "SMALLINT")
+    @Column(columnDefinition = "TINYINT")
     private Integer level;
 
     @NotNull
-    @Column(columnDefinition = "SMALLINT")
+    @Column(columnDefinition = "TINYINT")
     private Integer classNumber;
 }
